requirements.txt
Flask
Flask-SocketIO
python-socketio
PyYAML
python-dotenv
cerebras-cloud-sdk
eventlet
Jinja2
itsdangerous
Werkzeug
MarkupSafe
google-search-results